refactor: move components to developer route
