--- packages/desktop/components/buttons/index.js
@@ -1,1 +1,0 @@
-export * from './popup-buttons'
