+++ packages/desktop/components/index.js
@@ -1,4 +1,3 @@
-export * from './buttons'
 export * from './drawers'
 export * from './filter'
 export * from './menus'
