+++ packages/desktop/views/dashboard/developer/components/CreateAliasButton.svelte
@@ -1,19 +1,17 @@
 <script lang="ts">
     import { OnboardingButton } from '@ui'
-
     import { localize } from '@core/i18n'
-
     import { openPopup, PopupId } from '@desktop/auxiliary/popup'
 
-    function onClick(): void {
+    function onCreateAliasClick(): void {
         openPopup({
-            id: PopupId.TestDeepLinkForm,
+            id: PopupId.AliasConfirmation,
         })
     }
 </script>
 
 <OnboardingButton
-    primaryText={localize('actions.testDeepLink')}
-    secondaryText={localize('general.testDeepLinkDescription')}
-    {onClick}
+    primaryText={localize('actions.createAlias')}
+    secondaryText={localize('actions.createAliasDescription')}
+    onClick={onCreateAliasClick}
 />
