+++ packages/desktop/views/dashboard/developer/components/MintNativeTokenButton.svelte
@@ -1,11 +1,9 @@
 <script lang="ts">
-    import { OnboardingButton } from '@ui'
-
     import { selectedAccount } from '@core/account/stores'
     import { localize } from '@core/i18n'
     import { resetMintTokenDetails } from '@core/wallet'
-
     import { closePopup, openPopup, PopupId } from '@desktop/auxiliary/popup'
+    import { OnboardingButton } from '@ui'
 
     $: hasAliases = $selectedAccount.balances?.aliases.length > 0
