+++ packages/desktop/views/dashboard/developer/components/MintNftButton.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
     import { OnboardingButton } from '@ui'
-
     import { localize } from '@core/i18n'
     import { resetMintNftDetails } from '@core/wallet'
     import { openPopup, PopupId } from '@desktop/auxiliary/popup'
